feat: adds lessons crud

+++ src/application/modules/chapter/use-cases/delete-chapter/delete-chapter-use-case.ts
@@ -1,4 +1,4 @@
-import { ChangeChaptersCountUseCase } from '@/application/modules/unit/use-cases/change-chapters-count/change-chapters-count-use-case';
+import { UnitRepository } from '@/application/modules/unit/repositories/unit-repository';
 import { InternalServerHTTPError } from '@/application/shared/http/errors/internal-server-http-error';
 import { NotFoundHTTPError } from '@/application/shared/http/errors/not-found-http-error';
 import { IUseCase } from '@/application/shared/http/interfaces/use-case';
@@ -13,7 +13,7 @@ type IOutput = void;
 export class DeleteChapterUseCase implements IUseCase<IInput, IOutput> {
   constructor(
     private readonly chapterRepo: ChapterRepository,
-    private readonly changeChaptersCountUseCase: ChangeChaptersCountUseCase,
+    private readonly unitRepo: UnitRepository,
   ) {}
 
   async execute({ chapterId }: IInput): Promise<void> {
@@ -23,10 +23,16 @@ export class DeleteChapterUseCase implements IUseCase<IInput, IOutput> {
       throw new NotFoundHTTPError('Capítulo não encontrada');
     }
 
+    const unit = await this.unitRepo.getUnit(chapter.unitId);
+
+    if (!unit) {
+      throw new NotFoundHTTPError('Unidade não encontrada');
+    }
+
     try {
       await this.chapterRepo.deleteChapter(chapterId);
 
-      await this.changeChaptersCountUseCase.execute({ operation: 'DECREMENT', unitId: chapter.unitId });
+      await this.unitRepo.changeChaptersCount(unit, 'DECREMENT');
     } catch {
       throw new InternalServerHTTPError('Ocorreu um erro ao excluir capítulo');
     }
